LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-import-blob-url.any-expected.txt

Blocked access to external URL https://www1.localhost:9443/workers/modules/resources/export-on-load-script.py
CONSOLE MESSAGE: Importing a module script failed.

Harness Error (FAIL), message = Script error.

PASS Static import.
FAIL Static import (cross-origin). promise_test: Unhandled rejection with value: "Importing a module script failed."
PASS Static import (redirect).
PASS Nested static import.
PASS Static import and then dynamic import.
PASS Dynamic import.
PASS Nested dynamic import.
PASS Dynamic import and then static import.
PASS eval(import()).
